Set abi3-py310 compatibility
When compiling the library, the resulting .whl will only be
compatible against the same exact python version it was
compiled against.

By specifying an abi minimum compatibility we make the library
compatible with multiple python version.

abi3-py310 for example makes the resulting wheel compatible
with any python version higher than 3.10

This change is required to make the library easier to compile
and compatible with ooniprobe

See:
- https://pyo3.rs/main/features.html#abi3
- https://pyo3.rs/main/building-and-distribution#py_limited_apiabi3